+++ amazon_creatorsapi/core/constants.py
@@ -1,5 +1,6 @@
 """Constants for the Amazon Creators API."""
 
+DEFAULT_HOST = "https://creatorsapi.amazon"
 DEFAULT_THROTTLING = 1
 DEFAULT_TIMEOUT = 30.0